refactor: make a new class
