+++ common-tools/clas-logging/src/main/java/org/jlab/logging/DefaultLogger.java
@@ -15,45 +15,6 @@ import java.util.logging.Logger;
  * @author Jeremy McCormick, SLAC
  */
 public class DefaultLogger {
-    /**
-     * Configure a logger so INFO and below go to stdout, WARNING and above to stderr, with custom formatting.
-     */
-    public static void configureLogger(Logger logger) {
-        logger.setUseParentHandlers(false);
-        java.util.logging.Formatter formatter = new java.util.logging.SimpleFormatter() {
-            private static final String format = "NEW LOGGER: %5$s%6$s%n";
-            @Override
-            public synchronized String format(java.util.logging.LogRecord lr) {
-                return String.format(format,
-                        lr.getSourceClassName(),
-                        lr.getSourceMethodName(),
-                        lr.getLoggerName(),
-                        lr.getLevel().getLocalizedName(),
-                        lr.getMessage(),
-                        lr.getThrown() == null ? "" : lr.getThrown());
-            }
-        };
-        java.util.logging.Handler infoHandler = new java.util.logging.StreamHandler(System.out, formatter) {
-            @Override
-            public synchronized void publish(java.util.logging.LogRecord record) {
-                if (record.getLevel().intValue() <= Level.INFO.intValue()) {
-                    super.publish(record);
-                    flush();
-                }
-            }
-        };
-        java.util.logging.Handler errorHandler = new java.util.logging.StreamHandler(System.err, formatter) {
-            @Override
-            public synchronized void publish(java.util.logging.LogRecord record) {
-                if (record.getLevel().intValue() >= Level.WARNING.intValue()) {
-                    super.publish(record);
-                    flush();
-                }
-            }
-        };
-        logger.addHandler(infoHandler);
-        logger.addHandler(errorHandler);
-    }
 
     private static void init(String resource) {
         // FIXME:  document why this check is here:
@@ -66,8 +27,6 @@ public class DefaultLogger {
                 throw new RuntimeException("Initialization of default logging configuration failed.", e);
             }
         }
-        // Always apply custom handler after config is loaded
-        configureLogger(Logger.getLogger("") /* root logger */);
     }
 
     private static void init(Level level){
